requirements.txt
httpx
python-dotenv
loguru
